Updated code
Made code more readable in navigation_fixedwing.c
Increased multiply LC to int32

+++ src/main/programming/logic_condition.c
@@ -183,7 +183,7 @@ static int logicConditionCompute(
             break;
 
         case LOGIC_CONDITION_MUL:
-            return constrain(operandA * operandB, INT16_MIN, INT16_MAX);
+            return constrain(operandA * operandB, INT32_MIN, INT32_MAX);
             break;
 
         case LOGIC_CONDITION_DIV:
